+++ lib/set-schema-now/set-schema-now.d.ts
@@ -1,2 +1,2 @@
 import * as dgraph from 'dgraph-js';
-export declare function XSetSchemaNow(schema: string, dgraphClient: dgraph.DgraphClient, _dgraph?: typeof dgraph): Promise<dgraph.Payload>;
+export declare function xSetSchemaNow(schema: string, dgraphClient: dgraph.DgraphClient, _dgraph?: typeof dgraph): Promise<dgraph.Payload>;