+++ lib/test-helpers/drob-db.d.ts
@@ -1,2 +1,2 @@
 import * as dgraph from 'dgraph-js';
-export declare function XDropDBNow(c: dgraph.DgraphClient, _dgraph?: typeof dgraph): Promise<dgraph.Payload>;
+export declare function xDropDBNow(c: dgraph.DgraphClient, _dgraph?: typeof dgraph): Promise<dgraph.Payload>;