+++ lib/upsert-now/upsert-fns/basic-equality-upsert-fn.d.ts
@@ -1,0 +1,2 @@
+import { IUpsertFnReturnValues } from '../upsert-now';
+export declare const basicEqualityUpsertFn: (searchPredicates: any) => (node: any) => IUpsertFnReturnValues;