+++ lib/upsert-now/upsert-now.d.ts
@@ -1,7 +1,11 @@
 import * as dgraph from 'dgraph-js';
-export declare function XUpsertNow(searchPredicates: string | string[], data: object[], dgraphClient: dgraph.DgraphClient, _dgraph?: any): Promise<string[]>;
-export declare function XUpsertNow(searchPredicates: string | string[], data: object, dgraphClient: dgraph.DgraphClient, _dgraph?: any): Promise<string>;
-export declare function buildUpsertQuery(_searchPredicates: string | string[], node: object): {
-    query: string;
-    searchValues: string[];
-};
+export interface INodeFoundFunction {
+    existingUid: string | null;
+    newNodeFn?: (node: any) => object;
+}
+export interface IUpsertFnReturnValues {
+    dgraphQuery: string;
+    nodeFoundFn: (queryResult: dgraph.Response) => INodeFoundFunction;
+}
+export declare function xUpsertNow(upsertFn: (input?: any) => IUpsertFnReturnValues, data: object[], dgraphClient: dgraph.DgraphClient, _dgraph?: any): Promise<string[]>;
+export declare function xUpsertNow(upsertFn: (input?: any) => IUpsertFnReturnValues, data: object, dgraphClient: dgraph.DgraphClient, _dgraph?: any): Promise<string>;