+++ src/set-json/set-json.ts
@@ -1,14 +1,14 @@
 import * as dgraph from 'dgraph-js'
 import {Mutation} from 'dgraph-js';
 
-export function XSetJSON(object: object, _dgraph=dgraph): Mutation  {
+export function xSetJSON(object: object, _dgraph=dgraph): Mutation  {
     const mu = new _dgraph.Mutation();
     mu.setSetJson(object);
     return mu;
 }
 
-export function XSetJSONCommitNow(object: object, _dgraph=dgraph): Mutation {
-   const mu = XSetJSON(object, _dgraph);
+export function xSetJSONCommitNow(object: object, _dgraph=dgraph): Mutation {
+   const mu = xSetJSON(object, _dgraph);
    mu.setCommitNow(true);
    return mu
 }