+++ src/test-helpers/drob-db.ts
@@ -1,6 +1,6 @@
 import * as dgraph from 'dgraph-js';
 
-export function XDropDBNow(c: dgraph.DgraphClient, _dgraph = dgraph): Promise<dgraph.Payload> {
+export function xDropDBNow(c: dgraph.DgraphClient, _dgraph = dgraph): Promise<dgraph.Payload> {
     const op = new _dgraph.Operation();
     op.setDropAll(true);
     return c.alter(op);