+++ src/test-helpers/setup.ts
@@ -1,9 +1,9 @@
 import * as dgraph from 'dgraph-js';
 import * as messages from "dgraph-js/generated/api_pb";
-import {ICreateDgraphClientConfig, XCreateDgraphClient} from '../create-client/create-dgraph-client';
-import {XSetSchemaNow} from '..';
-import {XDropDBNow} from './drob-db';
-import {XSetJSONNow} from '../';
+import {ICreateDgraphClientConfig, xCreateDgraphClient} from '../create-client/create-dgraph-client';
+import {xSetSchemaNow} from '..';
+import {xDropDBNow} from './drob-db';
+import {xSetJSONNow} from '../';
 
 export interface ISetupReturnValue {
     dgraphClient: dgraph.DgraphClient,
@@ -11,7 +11,7 @@ export interface ISetupReturnValue {
     result?: messages.Assigned
 }
 
-export async function XSetupForTestNow(config?: ICreateDgraphClientConfig, _XCreateDgraphClient=XCreateDgraphClient, _drop=XDropDBNow): Promise<ISetupReturnValue> {
+export async function xSetupForTestNow(config?: ICreateDgraphClientConfig, _xCreateDgraphClient=xCreateDgraphClient, _drop=xDropDBNow): Promise<ISetupReturnValue> {
 
     const defaults = {
         port: 9081,
@@ -21,7 +21,7 @@ export async function XSetupForTestNow(config?: ICreateDgraphClientConfig, _XCre
 
     const testConfig: ICreateDgraphClientConfig = Object.assign(defaults, config);
 
-    const {dgraphClient, dgraphClientStub} = _XCreateDgraphClient(testConfig);
+    const {dgraphClient, dgraphClientStub} = _xCreateDgraphClient(testConfig);
     await _drop(dgraphClient);
     return {dgraphClient, dgraphClientStub};
 }
@@ -32,7 +32,7 @@ export interface ISetupWithParams {
     debugDgraphClient?: boolean
 }
 
-export async function XSetupWithSchemaDataNow(params: ISetupWithParams): Promise<ISetupReturnValue> {
+export async function xSetupWithSchemaDataNow(params: ISetupWithParams): Promise<ISetupReturnValue> {
     const defaultParams: ISetupWithParams = {
         schema: null,
         data: null,
@@ -41,12 +41,12 @@ export async function XSetupWithSchemaDataNow(params: ISetupWithParams): Promise
 
     const {schema, data, debugDgraphClient } = Object.assign(defaultParams, params);
 
-    const {dgraphClient, dgraphClientStub} = await XSetupForTestNow({debug: debugDgraphClient});
-    await XSetSchemaNow(schema, dgraphClient);
+    const {dgraphClient, dgraphClientStub} = await xSetupForTestNow({debug: debugDgraphClient});
+    await xSetSchemaNow(schema, dgraphClient);
 
     let result = null;
     if (data) {
-        result = await XSetJSONNow(data, dgraphClient)
+        result = await xSetJSONNow(data, dgraphClient)
     }
 
     return {dgraphClient, dgraphClientStub, result}