+++ src/upsert-now/upsert-fns/basic-equality-upsert-fn.spec.ts
@@ -1,20 +1,23 @@
-import {buildUpsertQuery} from './upsert-now';
+import {basicEqualityUpsertFn} from './basic-equality-upsert-fn';
 
-describe('buildUpsertQuery', () => {
+// todo check other query return values
+describe('xbasicEqualityQueryFn', () => {
     it('should build a query when matching a single term', () => {
         const searchPredicate = 'name';
         const data = {
             name: 'Cameron',
             age: 35
         };
-        const {query} = buildUpsertQuery(searchPredicate, data);
+
+        const queryFn = basicEqualityUpsertFn(searchPredicate);
+        const {dgraphQuery} = queryFn(data);
 
         const expectedQuery = `{
             q(func: eq(name, "Cameron")) {
                 uid
             }
         }`;
-        expect(query).toEqual(expectedQuery)
+        expect(dgraphQuery).toEqual(expectedQuery)
     });
 
     it('should build a query when matching a two terms', () => {
@@ -24,7 +27,9 @@ describe('buildUpsertQuery', () => {
             email: 'cam@gmail.com',
             age: 35
         };
-        const {query} = buildUpsertQuery(searchPredicates, data);
+
+        const queryFn = basicEqualityUpsertFn(searchPredicates);
+        const {dgraphQuery} = queryFn(data);
 
         const expectedQuery = `
         {
@@ -34,7 +39,7 @@ describe('buildUpsertQuery', () => {
                 uid
             }
         }`;
-        expect(query).toEqual(expectedQuery)
+        expect(dgraphQuery).toEqual(expectedQuery)
     });
 
     it('should build a query when matching more than two terms', () => {
@@ -45,7 +50,9 @@ describe('buildUpsertQuery', () => {
             twitter: 'vespertilian',
             age: 35
         };
-        const {query} = buildUpsertQuery(searchPredicates, data);
+
+        const queryFn = basicEqualityUpsertFn(searchPredicates);
+        const {dgraphQuery} = queryFn(data);
 
         const expectedQuery = `
         {
@@ -56,7 +63,21 @@ describe('buildUpsertQuery', () => {
                 uid
             }
         }`;
-        expect(query).toEqual(expectedQuery)
-    })
+        expect(dgraphQuery).toEqual(expectedQuery)
+    });
 
+    it('should throw an error if the key predicate is not present on the object being used for the upsert', async() => {
+        const searchPredicates = ['name', 'email'];
+        const data = {
+            name: 'Cameron',
+        };
+
+        const queryFn = basicEqualityUpsertFn(searchPredicates);
+
+        expect(() => queryFn(data)).toThrowError(`
+        The search predicate/s must be a value on the object you are trying to persist.
+        
+        "email" does not exist on:
+        {"name":"Cameron"}`);
+    });
 });